pyproject.toml
[project]
name = "image-matching"
version = "0.1.0"
description = "Image matching for human faces"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clip",
    "deepface>=0.0.95",
    "facenet-pytorch>=2.5.3",
    "geopy>=2.4.1",
    "insightface>=0.7.3",
    "ipykernel>=6.30.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "onnxruntime>=1.22.1",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.2",
    "piexif>=1.1.3",
    "pillow>=11.3.0",
    "pillow-heif>=0.13.0",
    "qdrant-client>=1.15.1",
    "retina-face>=0.0.17",
    "tf-keras>=2.20.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.56.1",
]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }

